move gui utility methods to their own class
